+++ core/new-gui/src/app/dashboard/admin/component/admin-user.component.scss
@@ -6,7 +6,7 @@
   border-radius: 50px;
   color:white;
   display: block;
-  width: 80px;
+  width: 86px;
   text-align: center;
 }
 
@@ -20,4 +20,8 @@
 
 .c2{
   background-color:lightcoral;
+}
+
+.c3{
+  background-color:lightslategray;
 }
