+++ core/new-gui/src/app/dashboard/admin/service/admin-user.service.ts
@@ -7,6 +7,8 @@ export const USER_BASE_URL = `${AppSettings.getApiEndpoint()}/admin/user`;
 export const USER_LIST_URL = `${USER_BASE_URL}/list`;
 export const USER_UPDATE_URL = `${USER_BASE_URL}/update`;
 
+export const USER_DELETE_URL = `${USER_BASE_URL}/delete`;
+
 @Injectable({
   providedIn: "root",
 })
@@ -22,5 +24,9 @@ export class AdminUserService {
       uid: uid,
       role: role,
     });
+  }
+
+  public delete(uid: number): Observable<Response> {
+    return this.http.delete<Response>(`${USER_DELETE_URL}/` + uid);
   }
 }
